requirements.txt
customtkinter
mysql-connector-python
Pillow
sqlglot
matplotlib
networkx
requests
sqlparse